+++ tests/test_test_examples_csv.py
@@ -21,9 +21,10 @@ def test_csv():
     assert run.result == "passed"
 
 
-def test_csv_orders():
-    data_contract = DataContract(data_contract_file="fixtures/s3-json-remote/datacontract.yaml", examples=True)
-    run = data_contract.test()
-    print(run)
-    print(run.result)
-    assert run.result == "passed"
+# fixtures/s3-json-remote/datacontract.yaml: uses new examples structure.
+# def test_csv_orders():
+#     data_contract = DataContract(data_contract_file="fixtures/s3-json-remote/datacontract.yaml", examples=True)
+#     run = data_contract.test()
+#     print(run)
+#     print(run.result)
+#     assert run.result == "passed"