+++ tests/test_test_s3_json_remote.py
@@ -1,7 +1,3 @@
-import os
-
-import pytest
-
 from datacontract.data_contract import DataContract
 
 # logging.basicConfig(level=logging.INFO, force=True)
@@ -9,25 +5,13 @@ from datacontract.data_contract import DataContract
 datacontract = "fixtures/s3-json-remote/datacontract.yaml"
 
 
-# Disabled, as this test fails when another local s3 test runs, not clear why.
-# Maybe with env variables or the DuckDB connection...
-def _test_test_s3_json():
-    if "AWS_ACCESS_KEY_ID" in os.environ:
-        pytest.fail("Failed: AWS_ACCESS_KEY_ID is set, which you break this test")
-    if "AWS_SECRET_ACCESS_KEY" in os.environ:
-        pytest.fail("Failed: AWS_SECRET_ACCESS_KEY is set, which you break this test")
-    if "DATACONTRACT_S3_ACCESS_KEY_ID" in os.environ:
-        pytest.fail("Failed: DATACONTRACT_S3_ACCESS_KEY_ID is set, which you break this test")
-    if "DATACONTRACT_S3_SECRET_ACCESS_KEY" in os.environ:
-        pytest.fail("Failed: DATACONTRACT_S3_SECRET_ACCESS_KEY is set, which you break this test")
+def test_test_s3_json(monkeypatch):
+    monkeypatch.delenv("AWS_ACCESS_KEY_ID", raising=False)
+    monkeypatch.delenv("AWS_SECRET_ACCESS_KEY", raising=False)
+    monkeypatch.delenv("DATACONTRACT_S3_ACCESS_KEY_ID", raising=False)
+    monkeypatch.delenv("DATACONTRACT_S3_SECRET_ACCESS_KEY", raising=False)
+
     data_contract = DataContract(data_contract_file=datacontract)
-
-    # Get all environment variables as a dictionary
-    env_vars = os.environ
-
-    # Print each environment variable
-    for key, value in env_vars.items():
-        print(f"{key}: {value}")
 
     run = data_contract.test()